+++ doc/changes.rst
@@ -2,15 +2,43 @@
 desitarget Change Log
 =====================
 
-0.31.2 (unreleased)
+0.32.1 (unreleased)
 -------------------
 
-* Add ELG/LRG/QSO/STD selections for commissioning [`PR #519`_].
-* Add full set of columns to supplemental skies file [`PR #518`_]
-* Fix some corner cases when reading HEALPixel-split files [`PR #518`_]
+* Add info on versioning, main_cmx_or_sv to bitmask notebook [`PR #527`_]
+
+.. _`PR #527`: https://github.com/desihub/desitarget/pull/527
+
+0.32.0 (2019-08-07)
+-------------------
+
+* Add URAT catalog information [`PR #526`_]. Includes:
+    * New module to retrieve URAT data from Vizier and reformat it.
+    * Code to match RAs/Decs to URAT, as part of that new URAT module.
+    * Substitute URAT PMs for GFAs where Gaia has not yet measured PMs.
+* Update CMX and Main Survey target classes [`PR #525`_]. Includes:
+    * New `SV0_WD`, `SV0_STD_FAINT` target classes for commissioning.
+    * Mild updates to `SV0_BGS` and `SV0_MWS` for commissioning.
+    * New `BGS_FAINT_HIP` (high-priority BGS) Main Survey class.
+    * Explicit checking on `ASTROMETRIC_PARAMS_SOLVED` for MWS targets.
+    * Add 3-sigma parallax slop in `MWS_MAIN` survey target class.
+* Add ``OBSCONDITIONS`` to target files [`PR #523`_] Also includes:
+    * Split target files explicitly into bright and "graydark" surveys.
+    * Default to such a file-spilt for SV and Main (not for cmx).
+    * Adds an informational bit for supplemental sky locations.
+* Use ``MASKBITS`` instead of ``BRIGHTSTARINBLOB`` [`PR #521`_]. Also:
+    * Extra options and checks when making and vetting bundling scripts.
+    * Option to turn off commissioning QSO cuts to speed unit tests.
+* Add ELG/LRG/QSO/STD selection cuts for commissioning [`PR #519`_].
+* Add full set of columns to supplemental skies file [`PR #518`_].
+* Fix some corner cases when reading HEALPixel-split files [`PR #518`_].
 
 .. _`PR #518`: https://github.com/desihub/desitarget/pull/518
 .. _`PR #519`: https://github.com/desihub/desitarget/pull/519
+.. _`PR #521`: https://github.com/desihub/desitarget/pull/521
+.. _`PR #523`: https://github.com/desihub/desitarget/pull/523
+.. _`PR #525`: https://github.com/desihub/desitarget/pull/525
+.. _`PR #526`: https://github.com/desihub/desitarget/pull/526
 
 0.31.1 (2019-07-05)
 -------------------
@@ -22,7 +50,7 @@ desitarget Change Log
 0.31.0 (2019-06-30)
 -------------------
 
-* MASKBITS of BAILOUT for randoms when no file is found [`PR #515`_].
+* ``MASKBITS`` of ``BAILOUT`` for randoms when no file is found [`PR #515`_].
 * Near-trivial fix for an unintended change to the isELG API introduced in `PR
   #513`_ [`PR #514`_]. 
 * Preliminary ELG cuts for DR8 imaging for main and SV [`PR #513`_].